Foundation completed
The basics of the language is done as it is not able to do the following
correctly use static types and make sure the types are correct type
separated float, int, and strings.
ability to declare and assign values.
for now, this will be all before the
better error handling.
This will be all before this weeks milestone