Bool, Print & Run
Added the ability to have bools to be stored and used.
Added the ability to have functions, for now all there is is print and run.
run is used to run code files such as py files in python, and print has not been implemented yet but will be soon.