Conditional and better math
Fixed up status to work correctly and now has built in ways to drain energy, thirst, and hunger.
polished how errors and warning are displayed to be easier and with the use of methods.
added operators such as greater than or equal which will result in true or false vaues.
Added the ability for if statements to work, with some small things needed to be fixed in the future, also added foundation for other statement's to be made, such as else, while, and more.
Also have list working, with the ability to extract data form them, will soon add methods such as append and delete.